DELETE

+++ controller/product.js
@@ -9,7 +9,7 @@ const { validationResult } = require("express-validator");
 
 exports.getProducts = async (req, res, next) => {
     const currentPage = req.query.currentPage || 1;
-    const perPage = req.query.perPage || 10;
+    const perPage = req.query.perPage || 17;
 
     const products = await Product.find()
         .skip((currentPage - 1) * perPage)
@@ -24,19 +24,28 @@ exports.getProducts = async (req, res, next) => {
         message: "Products Fetched",
         products: products,
         totalPages: totalPages,
+        status: "success",
+    });
+};
+
+exports.getAdminProducts = async (req, res, next) => {
+    const userProducts = await User.findById(req.userId).populate("products");
+    res.status(200).json({
+        products: userProducts.products,
+        status: "success",
     });
 };
 
 //  --------------------------------------------------------------------------------------------------------------  //
 
 exports.getDetails = async (req, res, next) => {
-    const prodId = req.query.prodId;
+    const prodId = req.params.prodId;
     const product = await Product.findById(prodId);
 
     if (!product)
         return next(createError.InternalServerError("Cannot reach Database"));
 
-    res.status(200).json(product);
+    res.status(200).json({ product: product, status: "success" });
 };
 
 //  --------------------------------------------------------------------------------------------------------------  //
@@ -53,7 +62,8 @@ exports.addProduct = async (req, res, next) => {
     const description = req.body.description;
     const quantity = req.body.quantity;
 
-    if (!req.files) return next(createError.UnsupportedMediaType("Invalid File Type"));
+    if (!req.files)
+        return next(createError.UnsupportedMediaType("Invalid File Type"));
 
     const product = new Product({
         title: title,
@@ -87,7 +97,8 @@ exports.getEditProduct = async (req, res, next) => {
 
 exports.editProduct = async (req, res, next) => {
     const error = validationResult(req);
-    if (!error.isEmpty()) return next(createError.Forbidden(error.array()[0].msg));
+    if (!error.isEmpty())
+        return next(createError.Forbidden(error.array()[0].msg));
 
     const prodId = req.query.prodId;
     const title = req.body.title;
@@ -118,8 +129,19 @@ exports.editProduct = async (req, res, next) => {
 exports.deleteProduct = async (req, res, next) => {
     const prodId = req.params.prodId;
     const product = await Product.findById(prodId);
+    if (!product) {
+        return createError.NotFound("Product Not Found");
+    }
     await imageHandler.unlinkImage(product.imgPath);
     await imageHandler.unlinkImage(product.previewImage);
     await Product.findByIdAndRemove(prodId);
-    res.status(200).json({ deleted: true });
+    const userProduct = await User.findById(req.userId);
+    userProduct.products.map((a) => {
+        if (a.toString() == prodId) {
+            userProduct.products.splice(userProduct.products.indexOf(a), 1);
+        }
+    });
+    const user = await userProduct.save();
+    res.status(200).json({ deleted: true, user: user });
 };
+
